# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Accès Direct Aide API",
    "version": "1.0.0",
    "description": "Documentation complète de l'API de la plateforme Accès Direct Aide — aide sociale, démarches, structures, rendez-vous professionnels et administration.",
    "contact": {
      "name": "Équipe AccesDirectAide",
      "url": "https://www.accesdirectaide.fr"
    }
  },
  "servers": [
    {
      "url": "https://www.accesdirectaide.fr/api",
      "description": "Production"
    },
    {
      "url": "http://localhost:5173/api",
      "description": "Développement local (Vite proxy)"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Développement local (API directe)"
    }
  ],
...
  "paths": {
...
    "/assistant/chat": {
      "post": {
        "tags": ["Assistant"],
        "summary": "Chat avec l'assistant IA",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Réponse de l'assistant (streaming SSE)" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatRequest": {
        "type": "object",
        "required": ["message"],
        "properties": {
          "message": { "type": "string" },
          "conversationId": { "type": "string", "format": "uuid" },
          "context": {
            "type": "object",
            "properties": {
              "departement": { "type": "string" },
              "situation": { "type": "string" }
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtenu via /auth/login ou /pro/auth/login"
      },
      "AdminToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-token",
        "description": "Token d'administration"
      },
      "CronSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "authorization",
        "description": "Bearer token cron (CRON_SECRET)"
      }
    }
  }
}